A driver for the BMP180 has been written
Functions for data acquisition, sensor initialization, and coefficient calculation for adequate pressure data reception have been added

+++ inc/sensor.h
@@ -16,7 +16,7 @@ struct sensor_t {
     sensor_address address;
     void* config_data;
     void (*init_fn)(void);
-    double (*get_data_fn)(void);
+    uint32_t (*get_data_fn)(void);
 };
 
 /*---------------------------------------------------------------------------*/
